// src/app/core/api/model/buildNamespaceResourcesResponseDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { QueuedBuildInfoDto } from './queuedBuildInfoDto';
import { BuildJobInfoDto } from './buildJobInfoDto';
import { BuildPodInfoDto } from './buildPodInfoDto';


export interface BuildNamespaceResourcesResponseDto { 
    namespace: string;
    /**
     * K8s Jobs currently running in flui-build
     */
    jobs: Array<BuildJobInfoDto>;
    /**
     * K8s Pods currently in flui-build
     */
    pods: Array<BuildPodInfoDto>;
    /**
     * Builds waiting in Bull queue (PENDING status, no K8s Job yet)
     */
    queuedBuilds: Array<QueuedBuildInfoDto>;
    /**
     * Sum of CPU requests across all running jobs (millicores)
     */
    totalCpuRequestMillicores: number;
    /**
     * Sum of memory requests across all running jobs (MiB)
     */
    totalMemoryRequestMiB: number;
}