// src/app/core/api/model/buildNamespaceCleanupResultDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BuildNamespaceCleanupResultDto { 
    deletedJobs: Array<string>;
    deletedPods: Array<string>;
    dryRun: boolean;
}