// src/app/core/api/model/buildJobInfoDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BuildJobInfoDto { 
    name: string;
    status: BuildJobInfoDto.StatusEnum;
    /**
     * Age in minutes
     */
    ageMinutes: number;
    buildId: string | null;
    appSlug: string | null;
    /**
     * e.g. bootstrap-runner for bootstrap jobs
     */
    purpose: string | null;
    cpuRequest: string;
    memoryRequest: string;
}
export namespace BuildJobInfoDto {
    export const StatusEnum = {
        Running: 'Running',
        Pending: 'Pending',
        Succeeded: 'Succeeded',
        Failed: 'Failed',
        Unknown: 'Unknown'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}