// src/app/core/api/model/queuedBuildInfoDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface QueuedBuildInfoDto { 
    buildId: string;
    applicationId: string;
    appSlug: string | null;
    branch: string;
    commitSha: string | null;
    /**
     * Age in minutes since the build was enqueued
     */
    ageMinutes: number;
    status: QueuedBuildInfoDto.StatusEnum;
}
export namespace QueuedBuildInfoDto {
    export const StatusEnum = {
        Pending: 'PENDING',
        Cloning: 'CLONING',
        Analyzing: 'ANALYZING',
        Building: 'BUILDING',
        Pushing: 'PUSHING'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}